Move block modules into their own subpackages
Give every module under scientific/blocks/ a subpackage, matching the layout
already used by neuron_sets/, stimuli/, synaptic_manipulations/ and
extracellular_locations/:

- recording.py -> recordings/{base,soma,ion_channel}.py, split along its three
  concerns: the Recording ABC, the soma voltage recordings, and the ion channel
  variable recordings.
- ion_channel_model.py -> ion_channel_model/ion_channel_model.py
- ion_channel_equations.py -> ion_channel_equations/ion_channel_equations.py
- afferent_synapses.py -> afferent_synapses/afferent_synapses.py

Pure code move, no behaviour change. The `type` discriminator used for
polymorphic (de)serialization is derived from the class name rather than the
module path, so previously serialized configs still deserialize unchanged.

Importers repointed: obi_one/__init__.py, the unions modules, the ion channel
modeling task (which imports ion_channel_equations as a module namespace), one
test, and three example notebooks.
