+++ obi_one/__init__.py
@@ -263,7 +263,7 @@ from obi_one.core.scan_generation import (
     GridScanGenerationTask,
     ScanGenerationTask,
 )
-from obi_one.scientific.blocks.afferent_synapses import (
+from obi_one.scientific.blocks.afferent_synapses.afferent_synapses import (
     AfferentSynapsesBlock,
     ClusteredPDSynapsesByCount,
     ClusteredPDSynapsesByMaxDistance,
@@ -355,8 +355,8 @@ from obi_one.scientific.blocks.neuron_sets.specific import (
     AllPopulationNeurons,
     AllVirtualNeurons,
 )
-from obi_one.scientific.blocks.recording import (
-    Recording,
+from obi_one.scientific.blocks.recordings.base import Recording
+from obi_one.scientific.blocks.recordings.soma import (
     SomaVoltageRecording,
     TimeWindowSomaVoltageRecording,
 )
