+++ obi_one/scientific/tasks/ion_channel_modeling.py
@@ -22,7 +22,9 @@ from obi_one.core.schema import SchemaKey
 from obi_one.core.serialization_constants import COORDINATE_CONFIG_FILENAME, SCAN_CONFIG_FILENAME
 from obi_one.core.single import SingleConfigMixin
 from obi_one.core.task import Task
-from obi_one.scientific.blocks import ion_channel_equations as equations_module
+from obi_one.scientific.blocks.ion_channel_equations import (
+    ion_channel_equations as equations_module,
+)
 from obi_one.scientific.from_id.ion_channel_recording_from_id import IonChannelRecordingFromID
 
 L = logging.getLogger(__name__)
